Refactor separate par and grid
