+++ Makefile
@@ -3,3 +3,4 @@ all:
 	cargo test --all-features
 	cargo test --no-default-features
 	cargo test --release
+	cargo package --allow-dirty
